// apps/web/src/api/generated.ts
export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderStatus = {
  ordered: "ordered",
  cooking: "cooking",
  delivered: "delivered",
} as const;

export interface Order {
  id: string;
  orderNumber: number;
  tableNumber: number;
  itemName: string;
  quantity: number;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}